Fix json serde issue (#141)
Use Double for temperatureOffset instead of Integer
That way it allows values such as -0.5 (as provided by Rika)

+++ rika-firenet/src/main/java/dev/cookiecode/rika2mqtt/rika/firenet/model/Controls.java
@@ -64,7 +64,7 @@ public class Controls {
   private Integer convectionFan2Area;
   private Boolean frostProtectionActive;
   private Integer frostProtectionTemperature;
-  private Integer temperatureOffset;
+  private Double temperatureOffset;
 
   @SerializedName("RoomPowerRequest") // for coherence (the rest of the api is using camelCase)
   private Integer roomPowerRequest;
